refactor login page, login component, api calls, file structure
