+++ src/Pages/AuthPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Login from '../Components/Login'
+
+const AuthPage = () => {
+    return (
+        <div>
+            <Login />
+        </div>
+    )
+}
+
+export default AuthPage
